seventh commit

+++ src/main/java/com/task9_springsecurity/model/Role.java
@@ -40,7 +40,8 @@ public class Role implements GrantedAuthority {
 
 
     public Integer getId() { return id; }
-    public void setId(Long id) { this.id = Math.toIntExact(id); }
+    public void setId(Integer id) { this.id = id; }
+
 
     public String getName() { return name; }
     public void setName(String name) { this.name = name; }
