+++ src/main/java/com/task9_springsecurity/model/User.java
@@ -51,7 +51,7 @@ public class User implements UserDetails {
 
     // getters & setters
     public Integer getId() { return id; }
-    public void setId(Long id) { this.id = Math.toIntExact(id); }
+    public void setId(Integer id) { this.id = id; }
 
     public String getFirstName() { return firstName; }
     public void setFirstName(String firstName) { this.firstName = firstName; }
